[swa-personal] set scratch buffer to markdown-mode
Over time, I've realized that most composition of emails or JIRA
tickets or JIRA comments or Confluence documents, should be inside the
respective org-mode headings.

Especially, because, that plays nicely with exporting to a file, since
the org-mode heading is in a file, and the auto-generated file will
use the same file name with different extension. Contrast that with
exporting from scratch buffer, where I have to provide the file name
every time.

So, what /should/ I use the scratch buffer for? For quick messages for
Slack, IRC, etc. That means markdown-mode would be preferable for that
purpose.